fixed bug - add/remove favouriteMovie from DetailFragment, and moved some files
