+++ Movies/app/src/main/java/spacekotlin/vaniukova/movies/MainActivity.kt
@@ -1,10 +1,11 @@
 package spacekotlin.vaniukova.movies
 
-import android.content.SharedPreferences
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import spacekotlin.vaniukova.movies.Navigator
+import spacekotlin.vaniukova.movies.R
 import spacekotlin.vaniukova.movies.databinding.ActivityMainBinding
 import spacekotlin.vaniukova.movies.favourite_movies.FavouriteMoviesFragment
 import spacekotlin.vaniukova.movies.movie_list_search.ListSearchMovieFragment
@@ -13,8 +14,6 @@ import spacekotlin.vaniukova.movies.top_movies_list.TopMoviesFragment
 class MainActivity : AppCompatActivity(R.layout.activity_main), Navigator {
 
     private lateinit var binding: ActivityMainBinding
-
-    val prefs: SharedPreferences? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
