+++ Movies/app/src/main/java/spacekotlin/vaniukova/movies/MovieApp.kt
@@ -8,6 +8,5 @@ class MovieApp: Application() {
     override fun onCreate() {
         super.onCreate()
         Database.init(this)
-        //AppPreferences.init(this)
     }
 }
