+++ Movies/app/src/main/java/spacekotlin/vaniukova/movies/top_movies_list/TopMoviesFragment.kt
@@ -15,7 +15,7 @@ import spacekotlin.vaniukova.movies.data.db.models.TopMovieDB
 import spacekotlin.vaniukova.movies.databinding.FragmentTopMoviesBinding
 import spacekotlin.vaniukova.movies.detail_movie.DetailFragment
 import spacekotlin.vaniukova.movies.movie.Movie
-import spacekotlin.vaniukova.movies.movie.MovieListAdapter
+import spacekotlin.vaniukova.movies.movie.adapter.MovieListAdapter
 import spacekotlin.vaniukova.movies.utils.autoCleared
 
 class TopMoviesFragment : Fragment(R.layout.fragment_top_movies) {
